Struct varisat_formula::cnf::CnfFormula

pub struct CnfFormula {
    var_count: usize,
    literals: Vec<Lit>,
    clause_ranges: Vec<Range<usize>>,
}

A formula in conjunctive normal form (CNF).

Equivalent to Vec<Vec> but more efficient as it uses a single buffer for all literals.

Fields
var_count: usizeliterals: Vec<Lit>clause_ranges: Vec<Range<usize>>
Methods

impl CnfFormula

pub fn new() -> CnfFormula

Create an empty CNF formula.

pub fn var_count(&self) -> usize

Number of variables in the formula.

This also counts missing variables if a variable with a higher index is present. A vector of this length can be indexed with the variable indices present.

pub fn set_var_count(&mut self, count: usize)

Increase the number of variables in the formula.

If the parameter is less than the current variable count do nothing.

pub fn len(&self) -> usize

Number of clauses in the formula.

pub fn iter(&self) -> impl Iterator<Item = &[Lit]>

Iterator over all clauses.